fix: moved APIConfig from internal to core
