+++ core/api_config.go
@@ -1,4 +1,4 @@
-package internal
+package core
 
 import "net/http"
