+++ core/go.mod
@@ -1,0 +1,3 @@
+module github.com/redhat-developer/app-services-sdk-go/core
+
+go 1.13
